docs: organize SDK integration ladder
