+++ package.json
@@ -89,9 +89,9 @@
     "example:sdk:add-tool": "tsx --no-cache examples/sdk/02-add-a-tool.ts",
     "example:sdk:add-mcp": "tsx --no-cache examples/sdk/03-add-an-mcp-server.ts",
     "example:sdk:custom-output": "tsx --no-cache examples/sdk/04-custom-output.ts",
-    "example:sdk:hosted-agent": "tsx --no-cache examples/sdk/hosted-agent/service-example.ts",
-    "example:sdk:hosted-api": "tsx --no-cache examples/sdk/hosted-agent/http-api-example.ts",
-    "example:sdk:browser-client": "tsx --no-cache examples/sdk/hosted-agent/browser-run-example.ts",
+    "example:sdk:hosted-agent": "tsx --no-cache examples/sdk/05-hosted-agent/01-hosted-service/run.ts",
+    "example:sdk:hosted-api": "tsx --no-cache examples/sdk/05-hosted-agent/02-http-sse-api/server.ts",
+    "example:sdk:browser-client": "tsx --no-cache examples/sdk/05-hosted-agent/03-browser-client/run.ts",
     "example:conversation-engine": "tsx --no-cache examples/conversation-engine.ts",
     "example:programmatic": "tsx --no-cache examples/programmatic-loop.ts",
     "example:heartbeat": "tsx --no-cache examples/heartbeat.ts",
